+++ Package.swift
@@ -18,6 +18,7 @@ let package = Package(
             linkerSettings: [
                 .linkedFramework("AppKit"),
                 .linkedFramework("Security"),
+                .linkedFramework("CryptoKit"),
                 .linkedLibrary("sqlite3")
             ]
         )